More elaborate version of demo game
Creating a separate branch for a version of the demo game which starts with a file system prompt. Since this obfuscates the usage of the game engine, I'm keeping it separate from the main branch. Although there is now redundant code, I think it's best to keep it that way so that the text-engine (index.js) in this branch can be easily upgraded with changes made to the master branch.